requirements.txt
Flask==3.0.3
numpy>=1.26.4
scipy>=1.10.1
scikit-learn>=1.4.2
matplotlib>=3.7.1
pandas>=2.2.2
tensorflow>=2.16.1
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
gast==0.2.2
gevent==20.5.0
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==22.0.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==3.3.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==1.1.1
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==10.3.0
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.16.2
tensorflow-estimator==2.15.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==3.0.2
wrapt==1.12.1
zipp==3.1.0
